+++ package.json
@@ -48,7 +48,7 @@
   },
   "lint-staged": {
     "*.{js, jsx, css, json}": [
-      "yarn run lint:fix",
+      "npm run lint:fix",
       "pretty-quick --staged",
       "git add"
     ]
