+++ src/components/App/App.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import HomePage from '../../pages/Home';
+
+function App() {
+  return (
+    <HomePage />
+  );
+}
+
+export default App;
