+++ src/components/App/index.js
@@ -1,1 +1,1 @@
-export { default } from './App.component';
+export { default } from './App';
