--- src/components/Layout/Layout.component.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import './Layout.styles.css';
-
-function Layout({ children }) {
-  return <main className="container">{children}</main>;
-}
-
-export default Layout;
