--- src/components/Private/index.js
@@ -1,1 +1,0 @@
-export { default } from './Private.component';
