+++ src/pages/Home/Home.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+function HomePage() {
+  return (
+    <section>
+    </section>
+  );
+}
+
+export default HomePage;
