--- src/pages/Home/Home.styles.css
@@ -1,8 +1,0 @@
-.homepage {
-  text-align: center;
-}
-
-.homepage h1 {
-  font-size: 3rem;
-  letter-spacing: -2px;
-}
