+++ src/pages/Home/index.js
@@ -1,1 +1,1 @@
-export { default } from './Home.page';
+export { default } from './Home';
