--- src/providers/Auth/index.js
@@ -1,1 +1,0 @@
-export { default, useAuth } from './Auth.provider';
