--- src/utils/constants.js
@@ -1,3 +1,0 @@
-const AUTH_STORAGE_KEY = 'wa_cert_authenticated';
-
-export { AUTH_STORAGE_KEY };
