--- src/utils/fns.js
@@ -1,5 +1,0 @@
-function random(limit) {
-  return Math.floor(Math.random() * limit);
-}
-
-export { random };
